requirements.txt
google.colab
matplotlib.pyplot
numpy
opencv-python
pillow
seaborn
tk
ultralytics
